improve motd.sh, add mic permission setup, update appstore setup
for motd.sh now use better smaller icons aand make the storage uses bar auto increase based on screen width.
a lot of people don't understand how to give mic permission to termux so i add a option which will prompt for mic permission based on termux-api is installed or not.
in appstore add installing a specific version so add that here to it will help me push all   more and more changes to the src branch